// src/features/categories/types/category.types.ts
export interface Category {
  _id: string;
  name: string;
  slug: string;
  image: string;
  createdAt: string;
  updatedAt: string;
}

export interface CategoryMetadata {
  currentPage: number;
  numberOfPages: number;
  limit: number;
}

export interface SingleCategoryResponse {
  data: Category;
}

export interface CategoryResponse {
  results: number;
  metadata: CategoryMetadata;
  data: Category[];
}